last update bevore final pre study
- changed score to float, updated highscore accordingly
- changed ring score from 1 to 0.2
- changed pov score from 5 to 1
- changed Task Pool Generator Method for Secondary Task Position from Take from Pool to Random Range
- Changed MaxMainTasksBeforeSecondary to 6
- Changed PovRepetitions to 3
- increased Pov trigger radius for selection
- changed Task Spawn Debug messages
- changed study finished text
- changed smell middle duration from 0.8 to 1s
- removed fade outs during study
- increased player speed from 20 to 25
- remastered bg music and added normalization

+++ Assets/Scripts/Player/ScoreDisplayControl.cs
@@ -3,7 +3,7 @@
 public class ScoreDisplayControl : MonoBehaviour
 {
 	private TextMesh scoreText;
-	private int score;
+	private float score;
 
 	//float timer;
 
@@ -25,27 +25,28 @@ public class ScoreDisplayControl : MonoBehaviour
 		}*/
 	}
 
-	public void SetScore(int score)
+	public void SetScore(float score)
 	{
 		this.score = score;
 		UpdateText();
 	}
 
-	public void AddScore(int score)
+	public void AddScore(float score)
 	{
 		this.score += score;
 	    if (score < 0) this.score = 0;
 		UpdateText();
 	}
 
-	public int GetScore()
+	public float GetScore()
 	{
 		return score;
 	}
 
 	private void UpdateText()
 	{
-		string scoreString = score < 10 ? "0" + score : "" + score;
+
+		string scoreString = score < 10 ? "0" + (int)score : "" + (int)score;
 
 		scoreText.text = scoreString;
 	}
